# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Prospect Finder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Prospect {
  id: number;
  name: string;
  position?: string;
  department?: string;
  company: string;
  country?: string;
  city?: string;
  email?: string;
  phone?: string;
  linkedinUrl?: string;
  seniority?: string;
  industry?: string;
  notes?: string;
  createdAt?: string;
}

export interface ProspectSearchResult {
  prospects: Prospect[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface SaveProspectRequest {
  prospectId: number;
  listName?: string;
  notes?: string;
}

export interface SavedProspect {
  id: number;
  prospectId: number;
  listName?: string;
  notes?: string;
  prospect: Prospect;
  savedAt?: string;
  pipelineStage?: string | null;
  lastContactedAt?: string | null;
  nextActionAt?: string | null;
  pipelineNotes?: string | null;
  cadenceStep?: number | null;
}

export type StatsResponseTopCountriesItem = {
  country: string;
  count: number;
};

export type StatsResponseTopIndustriesItem = {
  industry: string;
  count: number;
};

export interface StatsResponse {
  totalProspects: number;
  totalCompanies: number;
  totalCountries: number;
  topCountries: StatsResponseTopCountriesItem[];
  topIndustries: StatsResponseTopIndustriesItem[];
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface SocialProfile {
  platform: string;
  url: string;
  username?: string;
}

export interface PressItem {
  title: string;
  url: string;
  source?: string;
  date?: string;
  snippet?: string;
}

export type EnrichmentResultConfidence =
  (typeof EnrichmentResultConfidence)[keyof typeof EnrichmentResultConfidence];

export const EnrichmentResultConfidence = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface EnrichmentResult {
  prospectId: number;
  email?: string;
  alternativeEmails?: string[];
  phone?: string;
  linkedinUrl?: string;
  socialProfiles: SocialProfile[];
  pressItems: PressItem[];
  summary?: string;
  confidence?: EnrichmentResultConfidence;
  enrichedAt?: string;
  sources: string[];
}

export type SearchProspectsParams = {
  /**
   * Search mode
   */
  mode: SearchProspectsMode;
  /**
   * Company name filter
   */
  company?: string;
  /**
   * Country filter
   */
  country?: string;
  /**
   * City filter
   */
  city?: string;
  /**
   * Job position/title filter
   */
  position?: string;
  /**
   * Department filter
   */
  department?: string;
  /**
   * Person name filter (for specific_person mode)
   */
  name?: string;
  /**
   * Page number
   */
  page?: number;
  /**
   * Results per page
   */
  pageSize?: number;
};

export type SearchProspectsMode =
  (typeof SearchProspectsMode)[keyof typeof SearchProspectsMode];

export const SearchProspectsMode = {
  key_people: "key_people",
  by_company: "by_company",
  specific_person: "specific_person",
} as const;