# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Prospect Finder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prospects/{id}/enrich:
    post:
      operationId: enrichProspect
      tags: [prospects]
      summary: Enrich prospect with web data
      description: Searches the web for public contact info, social profiles and press mentions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrichment results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrichmentResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /prospects/{id}/enrichment:
    get:
      operationId: getProspectEnrichment
      tags: [prospects]
      summary: Get stored enrichment data for a prospect
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrichment data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrichmentResult"
        "404":
          description: No enrichment data found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SocialProfile:
      type: object
      properties:
        platform:
          type: string
        url:
          type: string
        username:
          type: string
      required:
        - platform
        - url
    PressItem:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        source:
          type: string
        date:
          type: string
        snippet:
          type: string
      required:
        - title
        - url
    EnrichmentResult:
      type: object
      properties:
        prospectId:
          type: integer
        email:
          type: string
        alternativeEmails:
          type: array
          items:
            type: string
        phone:
          type: string
        linkedinUrl:
          type: string
        socialProfiles:
          type: array
          items:
            $ref: "#/components/schemas/SocialProfile"
        pressItems:
          type: array
          items:
            $ref: "#/components/schemas/PressItem"
        summary:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        enrichedAt:
          type: string
          format: date-time
        sources:
          type: array
          items:
            type: string
      required:
        - prospectId
        - socialProfiles
        - pressItems
        - sources