# lib/api-zod/src/generated/types/socialProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Prospect Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface SocialProfile {
  platform: string;
  url: string;
  username?: string;
}